pyproject.toml
[tool.poetry]
name = "autonomous-twitter-bot"
version = "0.1.0"
description = "Autonomous Twitter bot system for New Orleans/music/GenZ domain engagement"
authors = ["Aaron Feingold <aaronfeingold@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
tweepy = "^4.14.0"
openai = "^1.40.0"
anthropic = "^0.34.0"
pymongo = "^4.6.0"
psycopg2-binary = "^2.9.9"
asyncio = "^3.4.3"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"
asyncpg = "^0.29.0"
motor = "^3.3.0"
pgvector = "^0.2.4"
prometheus-client = "^0.19.0"
schedule = "^1.2.0"
numpy = "^1.26.0"
tiktoken = "^0.5.2"
httpx = "^0.25.0"
python-dateutil = "^2.8.2"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
isort = "^5.12.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true